// frontend/src/app/model/tarifauto.model.ts
export interface TarifAuto {
  
    id: number
    brand?: string;
    model?: string;
    registration?: string;
    local: string;
    zipcode: string;
    city: string;
    dateOfBirth: string;
    ageObtainingPermit: number;
    bonus: string;
    lastname: string;
    firstname: string;
    phone: string;
    email: string;
    
  }
  